modified:   README.md 	modified:   app/routes/admin.py 	modified:   app/routes/auth.py 	modified:   app/routes/cats.py 	modified:   app/routes/main.py 	new file:   app/services/__init__.py 	new file:   app/services/cat_service.py 	new file:   app/services/user_service.py 	modified:   run.py

+++ app/routes/admin.py
@@ -2,8 +2,7 @@
 from flask import Blueprint, render_template, redirect, url_for, request
 from flask_login import login_required, current_user
 import os
-from .. import db
-from ..models import Cat
+from ..services.cat_service import CatService
 from ..forms import CatForm
 
 bp = Blueprint('admin', __name__, url_prefix='/admin')
@@ -13,7 +12,7 @@ bp = Blueprint('admin', __name__, url_prefix='/admin')
 def cats():
     if not current_user.is_admin:
         return redirect(url_for('main.home'))
-    cats = Cat.query.order_by(Cat.created_at.desc()).all()
+    cats = CatService.get_all_cats()
     return render_template('admin_cats.html', cats=cats)
 
 @bp.route('/edit_cat/<int:cat_id>', methods=['GET', 'POST'])
@@ -22,7 +21,7 @@ def edit_cat(cat_id):
     if not current_user.is_admin:
         return redirect(url_for('main.home'))
     
-    cat = Cat.query.get(cat_id)
+    cat = CatService.get_cat(cat_id)
     if not cat:
         return redirect(url_for('admin.cats'))
     
@@ -33,16 +32,19 @@ def edit_cat(cat_id):
         form.description.data = cat.description
     
     if form.validate_on_submit():
+        update_data = {
+            'name': form.name.data,
+            'description': form.description.data
+        }
+        
         if form.image.data:
             if cat.image and os.path.exists(os.path.join(current_app.config['UPLOAD_FOLDER'], cat.image)):
                 os.remove(os.path.join(current_app.config['UPLOAD_FOLDER'], cat.image))
             filename = secure_filename(form.image.data.filename)
             form.image.data.save(os.path.join(current_app.config['UPLOAD_FOLDER'], filename))
-            cat.image = filename
+            update_data['image'] = filename
         
-        cat.name = form.name.data
-        cat.description = form.description.data
-        db.session.commit()
+        CatService.update_cat(cat_id, **update_data)
         return redirect(url_for('admin.cats'))
     
     return render_template('edit_cat.html',
@@ -56,11 +58,9 @@ def delete_cat(cat_id):
     if not current_user.is_admin:
         return redirect(url_for('main.home'))
     
-    cat = Cat.query.get(cat_id)
-    if cat:
-        if cat.image and os.path.exists(os.path.join(current_app.config['UPLOAD_FOLDER'], cat.image)):
-            os.remove(os.path.join(current_app.config['UPLOAD_FOLDER'], cat.image))
-        db.session.delete(cat)
-        db.session.commit()
+    cat = CatService.get_cat(cat_id)
+    if cat and cat.image and os.path.exists(os.path.join(current_app.config['UPLOAD_FOLDER'], cat.image)):
+        os.remove(os.path.join(current_app.config['UPLOAD_FOLDER'], cat.image))
     
+    CatService.delete_cat(cat_id)
     return redirect(url_for('admin.cats'))
